[PiranhaJava] Fail hard in the presence of configuration errors.
This makes PiranhaConfigurationException a runtime exception,
and gets rid of the `disabled` checker state. In general, it doesn't
make much sense to run Piranha with invalid configuration and
it's better to show the issue to users as visibly as possible,
since either way the tool behavior will be broken in this case.

This resolves #76 and #77.