[PiranhaJava] Update RELEASING.md instructions.
The new instructions ensure that `[PiranhaJava]` appears
in commits related to releasing a new PiranhaJava version.
Additionally, tags for PiranhaJava version X.Y.Z now
include `-java`.

This is also intended as a template for releasing instructions
for Piranha versions for other supported languages, in terms
of commit titles and git tags.