[PiranhaJava] Support array expressions in ExperimentTest annotations
This adds support for `@ExperimentTest` style annotations which
take an array expression rather than a single value to their
`treated` parameter.

Before this change, analyzing such annotation would result in an
NPE.

In terms of the semantics/intended refactoring, consider code like:

```
@ExperimentTest(treated = {XP_FLAG_1,XP_FLAG_2,XP_FLAG_3})
public void foo() {
  ...
}
```

And `XP_FLAG_1` being stale.

If the flag is treated, Piranha should remove it from the list but
keep the annotation.

If the flag is in control, the entire test should be removed.

Additionally, this code:

```
@ExperimentTest(treated = {XP_FLAG_1})
public void foo() {
  ...
}
```

Should be treated as equivalent to:

```
@ExperimentTest(treated = XP_FLAG_1)
public void foo() {
  ...
}
```