[PiranhaJava] Switch Java CI to GitHub Actions. (#125)
Travis CI is about to be sunset for Uber OSS projects, we should switch PiranhaJava and PiranhaSwift CI builds to use GitHub Actions, just as PiranhaJS already does (see also #124). This PR accounts for the PiranhaJava portion.

This required a number of changes to our Gradle build. Most importantly, a long pending upgrade from Gradle 4.9 to 6.8.3, but also configuration to run the sample project on JDK11, fixes to our code coverage reporting, and an update to our Google Java Format plugin.