[PiranhaJava] Remove configured test methods
This development allows the user of Piranha to specify a list of methods
where if a dynamic flag evaluation is in a certain argument position, we
remove that method.

This is useful for cleaning up `mock()` wrapper methods or `assert()`
methods in test code that are useless after a feature flag is removed.

Limitations:
  - This does not remove unused imports for the test methods because we
    have to account for cases of the test methods that will not be
    removed.