new line
Update tests

Add test case and do one more bug fix.

It is related to the current bug fix. To write a test for it, and make it pass I had to fix both. I was planning on doing it as a followup but I just did it in this commit.