Bug Fix/ Feature ? Allow performing consecutive scope level rules.  (#225)
Currently, when we say apply Rule 1 and then apply Rule 2 in the class scope of Rule 1, we assume that Rule 1 is applied to a node which is a child node of the Class (i.e. Class node should be ancestor of the node changed by Rule 1). But it could happen that, Rule 1 changes the class node itself.
For such case, we now also match the scope query against the changed node itself and its ancestors.
Earlier we matched scope query only against ancestors.