Bugfix/Cleanup: Piranha was escaping and unescaping new line characters when not required  (#224)
Sometimes we need to escape new line characters when it is part of the tree-sitter query. i.e. convert \n to \\n.
Problem is that, I was using the method substitute_tags for both substituting tags in a replacement pattern and in a tree-sitter query. There was a bug because of this.
To fix this bug, I basically added a boolean flag indicating wether the input is a tree-sitter query . This flag determines if a new line character is escaped in the substituted value.

# Bug Fix/ Feature ? Allow performing consecutive scope level rules.  (#225)

Currently, when we say apply Rule 1 and then apply Rule 2 in the class scope of Rule 1, we assume that Rule 1 is applied to a node which is a child node of the Class (i.e. Class node should be ancestor of the node changed by Rule 1). But it could happen that, Rule 1 changes the class node itself.
For such case, we now also match the scope query against the changed node itself and its ancestors.
Earlier we matched scope query only against ancestors.