Bug Fix:  Piranha was incorrectly identifying the changed node in the recently applied edit. (#223)
The problem was that, we were using the matched range of the current edit when determining the changed node. IT should have been the range of the replacement of the current edit. When I did that I realized that I did not require the current_match_range at all. So I consequently removed it.

## Bugfix/Cleanup: Piranha was escaping and unescaping new line characters when not required  (#224)

Sometimes we need to escape new line characters when it is part of the tree-sitter query. i.e. convert \n to \\n.
Problem is that, I was using the method substitute_tags for both substituting tags in a replacement pattern and in a tree-sitter query. There was a bug because of this.
To fix this bug, I basically added a boolean flag indicating wether the input is a tree-sitter query . This flag determines if a new line character is escaped in the substituted value.

## Bug Fix/ Feature ? Allow performing consecutive scope level rules.  (#225)

Currently, when we say apply Rule 1 and then apply Rule 2 in the class scope of Rule 1, we assume that Rule 1 is applied to a node which is a child node of the Class (i.e. Class node should be ancestor of the node changed by Rule 1). But it could happen that, Rule 1 changes the class node itself.
For such case, we now also match the scope query against the changed node itself and its ancestors.
Earlier we matched scope query only against ancestors.