Delete associated comments  (#237)
Fixes #236

We want to delete the comments associated to the deleted code elements across all languages

@lazaroclapp at a higher level, i have a question - should we guard this with a Piranha argument, like delete_associated_comments (that defaults true)? Or that doesn't seem necessary?

General idea:

Make comment identification configurable for each language (basically, add a predicate that checks if a particular node is a comment for a particular language) . Look at this new method is_comment
Whenever a rule deletes something (i.e. the replace string is empty ) -
a. check if the line from which the ast node was deleted contains comment nodes only . IF so delete the comment node.
b. Check if the line from which the ast node was deleted is blank and perform 2a for the row above, else perform 2a for row above it. and call (2) recursively for the above row.
Update
Deleting associated comment can be configured by two flags in piranha_arguments.toml - cleanup_comments_buffer (usize) and cleanup_comments (boolean). The former flag determines - how empty lines to look up in to find associated comment the later determines if comment cleanup should be enabled or not (by default it is disabled). Example usage.

I know its expensive to do this comment analysis. Like unless we resort to regex based strategies (since we know the position), parsing the entire tree cannot be avoided. Its like do we want it flaky (current strategy which is expensive but more reliable) or more flaky (less reliable, less resource expensive) :|