Update on "Create a wrapper around strings carrying Tree-sitter queries"
**Problem:**
I was casually reviewing the code and I noticed that its very confusing when it comes to determining whether a string is carries a tree-sitter query or its just some other string. Currently, for this we completely rely upon the variable names. 
This can also lead to bugs. For instance, when we replace tags with values inside strings (we do a simple string replacement). However, when we replace tags with values inside tree-sitter-queries we need to escape it. 

**Solution:**
Create a wrapper _new type_ `TSQuery` around a String.
I also took this opportunity to extract  this `substitute_tag` behavior into a `Instantiate` trait. 
I implemented it for `String`, `TSQuery` and `Rule`.

-----



[ghstack-poisoned]