Update base for Update on "Create a wrapper around strings carrying Tree-sitter queries"
**Problem:**
I was casually reviewing the code and I noticed that its very confusing when it comes to determining whether a string carries a tree-sitter query or not. Currently we rely upon the variable names 😢  This can also lead to 🐛 .
For instance, when we replace tags with values inside strings, we perform a simple string replacement.
However, when we replace tags with values inside tree-sitter-queries we need to escape it `\n` -> `\\n`. 

**Solution:**
Create a wrapper _new type_ `TSQuery` around a String.
I also took this opportunity to extract  this `substitute_tag` behavior into a `Instantiate` trait. 
I implemented it for `String`, `TSQuery, `Rule` and `Constraint`.

(Passing that `boolean` based on whether a string is a tree-sitter query or not was a bit "Risky"

-----



[ghstack-poisoned]