Update base for Update on "Fully qualify types inside the `piranha_rule` macro"
Fairly straightforward refactoring.  Using `HashSet<T>` requires `T: Hash` which is an unnecessary constraint. 
Let's not use `HashSet` when using `Vec` is **ok**.

------

I have updated this PR. I have undone the previous changes, and kept some random cleanups (but sensible) for sake of not tinkering with the stack.

The change in this PR, fully qualifies `HashSet` in the `piranha_rule!` macro

----



[ghstack-poisoned]