Bug/Feature: Handle syntactically incorrect input trees (#386)
Background: Currently Piranha accepts syntactically correct code, and panics when an applied a rule creates syntactically incorrect code.

Feature: We want to allow Piranha to operate over partially correct ASTs when the user passes the option --allow-dirty-ast . We want this feature because it will (i) make code snippet mode better (not all code snippets are syntactically correct always , they might miss a brace here and there) (ii) allow us to operate over files with advanced language features that our grammar may not be supporting at the moment.

Implementation:

Add --alow-dirty-ast flag in PiranhaArguments
This flag ensures that a source code unit cannot be constructed if the AST is errorneous
Slightly tweak the core algorithm. Instead of panicking when an erroneous AST is produced, we now panic only of the number of errors in the AST have increased. This is a more general, than what we had earlier.
Test Plan: Added three test cases for:

Transform when input code is dirty, allow dirty is set and rule does not create more errors
Panic when input code is dirty, allow dirty is set and rule creates more errors
Panic when input code is dirty and allow dirty is not set