Gh/adarsh np/11/base - Var Inline 1/n (#456)
* Variable Inline [1/n]: Local variable inline

[ghstack-poisoned]

* Update on "Variable Inline [1/n]: Local variable inline"

This PR aims to solve Variable inlining for variables within the function scope. Please take a look at the tests to understand the various cases. 





[ghstack-poisoned]

* Update on "Variable Inline [1/n]: Local variable inline"

This PR aims to solve Variable inlining for variables within the function scope. Please take a look at the tests to understand the various cases. 





[ghstack-poisoned]

* Update on "Variable Inline [1/n]: Local variable inline"

This PR aims to solve Variable inlining for variables within the function scope. Please take a look at the tests to understand the various cases. 





[ghstack-poisoned]

* Update on "Variable Inline [1/n]: Local variable inline"

This PR aims to solve Variable inlining for variables within the function scope. Please take a look at the tests to understand the various cases. 





[ghstack-poisoned]