Update on "Allow user to include/exclude glob patterns from path_to_codebase"
Our users (internal) work upon large codebases. Moreover piranha currently allows user to set the path to the root of the codebase only. This implies that we end up `grepping` a lot of files that can be avoided (a huge performance setback). It also sometimes leads to un-needed transformations. 

In this PR, I introduce `--include` and `--exclude` flag that gives user fine-grained control over the specific directories/files they want to (and do not want to) analyze/transform.
Users can add specify this using glob patterns. 
I think the "interface" is in the spirit of cargo/pyright - where user provides path to root folder and define `include` / `exclude` files under it. 


**Test Plan : ** Adapted the `structural_find_and_replace` tests for this . 
In this test we basically look for all the usages of `exp.isTreated(STALE_FLAG)`.
Earlier, the directory structure was like : 
```
|- structural_find_and_replace 
  |- input 
      - SampleEnum.java 
      -  XPCleanupPositiveCases.java 
```

Now it is : 
```
|- structural_find_and_replace 
  |- input 
    |- folder_1 
      - SomeEnum.java 
      - SomeOtherClass.java 
    |- folder_2
      - SampleEnum.java 
      -  XPCleanupPositiveCases.java 
      |- folder_2_1
         - SomeClass.java
```
The PiranhaArguments for this test were updated to include `folder_2` but not `folder_2_1`(Therefore the number of usages remain the same) . Here we test that - Piranha only analyzes `folder_2` and not `folder_1` and `folder_2_1`.

We also test structural find replace via our Python API. This test was updated to include all files under `input` but not under `folder_2_1`. 

I think with this we cover all the scenarios : 
1. Test `include` feature only
2. Test `inlcude` and `exclude` feature
3. Test `exclude` feature only

-----




[ghstack-poisoned]