Build a tree-sitter playground with the grammars that Piranha uses (#736)
Piranha uses several grammar repositories with custom patches to support
the transformations. While
these patches are being upstreamed, there may be discrepancies between
the grammars in this
repository and the upstream grammars.

This PR adds a build script to

(1) instantiate the `index.html.template` (slightly modified from
[tree-sitter-cli-playground-html])
  and copy it to the given `dist` directory.

(2) clone the (custom) grammar repositories that we use in Piranha (by
parsing the `Cargo.toml`
file), check out the specific versions, and then build the WASM files
for the grammars and copy
  them to the `dist/assets` directory.

(3) the `dist` directory can be served by any web server (e.g., `python
-m http.server`).

We also add a GitHub Actions job to automatically deploy to GitHub Pages
whenever the grammars have been updated in Piranha (or the build script
/ GitHub Actions configuration chanegs).

This is tested in my fork:
https://yuxincs.github.io/piranha/tree-sitter-playground/

[tree-sitter-cli-playground-html]:
https://github.com/tree-sitter/tree-sitter/blob/eaa10b279f208b47f65e77833d65763f072f3030/crates/cli/src/playground.html#L13