Parse concrete syntax with a proper grammar (#741)
Concrete syntax has a hacky implementation based on string matching. If
we want to expand the grammar to support constraints and other
constructs, it's worthwhile to build a proper grammar. This will allow
us to expand it in the future to use things like:

- `where len(:[name]) > 2` for example to select nodes only with two or
more children;
- `where :[name] matches regex` with regex to add constraints on
template variables.

In this PR I essentially wrote the paper, but didn't touch the matching
algorithm. In summary:

- Added Pest grammar (concrete_syntax.pest)
- Added parser that converts strings to build an actual AST
- Moved the matching algorithm to
[interpreter.rs](https://github.com/danieltrt/piranha/blob/a993eb15fd9f2317f928c6a94c96577ffb73336d/src/models/concrete_syntax/interpreter.rs#L44)