Remove experimental directory (#745)
Per discussion w/ @danieltrt , the Piranha playground is no longer being
used and is safe to delete. If we need this in the future, we can always
reference it in git history, and we can properly refactor and maintain
it outside of experimental dir.

Removing the dir helps with maintainability.