Bug fix in parent scope rules (#778)
**Summary:**
There is a bug in piranha core that prevents parent rules from being
match only. Parent scope is forcibly applying edits without checking
whether the rule is match only. This PR fixes that


**Test plan:**
I added a test that fails without the fix to illustrate the problem, as
well as an E2E test